add shortcut for character editing from set goal
